==> Rmd/1_Preprocess.Rmd <==
---
title: "Google Analytics Customer: Preprocess"
author: "kotsubo takuto"
output: 
    html_document:
      md_extensions: -ascii_identifiers
      toc: true
      toc_depth: 3
      code_folding: hide
---

# Setting{.tabset .tabset-fade .tabset-pills}

## knitr option

```{r reset, include=FALSE}
# 初期化
rm(list = ls())
```

```{r set up, message=FALSE}
# set directory
setwd("~/Desktop/Google_Analytics_Kaggle/") 
# max.print 
options(max.print="200", digits=5)
# Global options
library(knitr)
opts_chunk$set(echo=TRUE,
               cache = FALSE,
	             prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

## Library package

```{r package, message=FALSE}
library(tidyverse)
library(readr) # for csv
library(summarytools) # summary easily for EDA
library(GGally) # ggpairs
library(skimr) 
library(janitor)
library(jsonlite)
```

## Load funciton

```{r}
source('~/Desktop/Google_Analytics_Kaggle/script/function.R') # for preprocessing
source('~/Desktop/Google_Analytics_Kaggle/script/makedummies.R') # for preprocessing
```

## Import csv

- Import raw data from `input/raw/{tarin|test}.csv`
- `device`, `geoNetwork`, `totals` and `trafficSource` are JSON style columns.

```{r cache=TRUE}
# read sample data
train_sample <- read_csv("~/Desktop/Google_Analytics_Kaggle/input/raw/train.csv",na = c("XNA","NA","","NaN","?"), n_max = 1000)
# extract columns class
col_vars_candidate <- train_sample %>% 
  map_chr(class)
col_vars <-  
      case_when(col_vars_candidate == "character" ~ "c",
                col_vars_candidate == "integer" ~ "i",
                col_vars_candidate == "numeric" ~ "d") %>% 
      purrr::set_names(names(col_vars_candidate)) %>% # set name by columns class
      as.list() %>% # list
      list_modify(date = "D", visitId = "c") %>% 
      flatten_chr() %>% # unlist
      str_c(collapse = "") # connect
# read raw data
train_raw <- read_csv("~/Desktop/Google_Analytics_Kaggle/input/raw/train.csv",
                      na = c("XNA","NA","","NaN","?"),
                      col_types = col_vars,
                      locale = readr::locale(date_format = "%Y%m%d"))
test_raw <- read_csv("~/Desktop/Google_Analytics_Kaggle/input/raw/test.csv",
                      na = c("XNA","NA","","NaN","?"),
                      col_types = col_vars,
                      locale = readr::locale(date_format = "%Y%m%d"))
# json columns
json_vars = c("device", "geoNetwork", "totals", "trafficSource")
# make a flatten data set
train_flat <- ReadJsonFile(train_raw, json_vars = json_vars)
test_flat <- ReadJsonFile(test_raw, json_vars = json_vars)
```

# Preprocess

- Remove columns for which information can not be obtained
- change character type and logical type to numeric type
- Replace NA with zero on target values
- Convert Unix time stamp to Date on `visitStartTime`
- Determine the local time on `visitStartTime`

## Remove columns for which information can not be obtained

- "not available in demo datase", "not set" and NA value are same?? 

```{r}
train_flat <-
  train_flat %>% 
  select(-c(sessionId,socialEngagementType,browserSize,browserVersion,operatingSystemVersion,
            mobileDeviceBranding,mobileDeviceModel,mobileDeviceMarketingName,mobileDeviceInfo,
            mobileInputSelector,flashVersion,language,screenColors,screenResolution,
            cityId,latitude,longitude,networkLocation,visits,campaignCode,adwordsClickInfo.criteriaParameters))
test_flat <-
  test_flat %>% 
  select(-c(sessionId,socialEngagementType,browserSize,browserVersion,operatingSystemVersion,
            mobileDeviceBranding,mobileDeviceModel,mobileDeviceMarketingName,mobileDeviceInfo,
            mobileInputSelector,flashVersion,language,screenColors,screenResolution,
            cityId,latitude,longitude,networkLocation,visits,adwordsClickInfo.criteriaParameters))
```

## change character type and logical type to numeric type

- 3つのIDはcharcter型で良いか? -> csv読み込み時,int型になるので変更しておく (`sessionId`以外)

```{r}
train_flat <-
  train_flat %>% 
  mutate_at(c("isMobile", "isTrueDirect","adwordsClickInfo.isVideoAd"), funs(as.numeric(.))) %>% 
  mutate_at(c("hits","pageviews","bounces","newVisits","transactionRevenue","adwordsClickInfo.page"), funs(as.numeric(.))) %>% 
  mutate_at(c("fullVisitorId","visitId"), funs(as.numeric(.)))
test_flat <-
  test_flat %>% 
  mutate_at(c("isMobile", "isTrueDirect","adwordsClickInfo.isVideoAd"), funs(as.numeric(.))) %>% 
  mutate_at(c("hits","pageviews","bounces","newVisits","adwordsClickInfo.page"), funs(as.numeric(.))) %>% 
  mutate_at(c("fullVisitorId","visitId"), funs(as.numeric(.)))
```

## Replace NA with zero on target values

```{r}
train_flat <- 
  train_flat %>% 
  mutate(transactionRevenue = ifelse(is.na(transactionRevenue),0,transactionRevenue)) 
```

## Convert Unix time stamp to Date on `visitStartTime`

- `date` is unnecessary because it is duplicated

```{r}
train_flat <- train_flat %>% 
  mutate(DateTimeSeattle = as.POSIXct(visitStartTime, origin='1970-01-01', tz = "US/Pacific")) %>% 
  select(-date)
test_flat <- test_flat %>% 
  mutate(DateTimeSeattle = as.POSIXct(visitStartTime, origin='1970-01-01', tz = "US/Pacific")) %>%   
  select(-date)
```

## Determine the local time

- continent: 大陸の中で最大数を持つ都市をtzに使用する
    - Africa: Africa/Tunis
    - Americas: America/Los_Angeles
    - Asia: Asia/Bangkok
    - Europe: Europe/London
    - Oceania: Australia/Sydney
- country: 
    - 首都をtime zone とする
- city: 
    - 大きな国(ロシア, 中国, アメリカ等)について個別に選択する
    - 巨大な都市については個別に入力
- POSIXct型は timezoneを指定する必要があるのでcharcter型にして, 読み込む時に戻す
- `visitStartTime` is unnecessary because it is duplicated

```{r }
train_flat <- 
  train_flat %>% 
  mutate(DateTimeLocal = # Continent
           case_when(
             continent == "Africa" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Tunis"),
             continent == "Americas" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Los_Angeles"),
             continent == "Asia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Bangkok"),
             continent == "Europe" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/London"),
             continent == "Oceania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Australia/Sydney"),
             TRUE ~ DateTimeSeattle
           )
         ) %>% 
  mutate(DateTimeLocal = # country
           case_when(
             country == "United States" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Los_Angeles"),
             country == "India" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Indian/Kerguelen"),
             country == "Canada" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Canada/Atlantic"),
             country == "Japan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Tokyo"),
             country == "Turkey" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Istanbul"),
             country == "Germany" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Berlin"),
             country == "Brazil" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Sao_Paulo"),
             country == "France" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Paris"),
             country == "Mexico" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Mexico_City"),
             country == "Taiwan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Taipei"),
             country == "Russia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Moscow"),
             country == "Spain" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Madrid"),
             country == "Netherlands" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Amsterdam"),
             country == "Italy" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Rome"),
             country == "Poland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Warsaw"),
             country == "Indonesia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jakarta"),
             country == "Philippines" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Manila"),
             country == "Singapore" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Singapore"),
             country == "Ireland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Dublin"),
             country == "Malaysia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Kuala_Lumpur"),
             country == "Romania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Bucharest"),
             country == "Ukraine" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Kiev"),
             country == "Israel" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jerusalem"),
             country == "Peru" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Lima"),
             country == "Sweden" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Stockholm"),
             country == "South Korea" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Seoul"),
             country == "Argentina" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Argentina/Buenos_Aires"),
             country == "Colombia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Bogota"),
             country == "Hong Kong" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Hong_Kong"),
             country == "Belgium" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Brussels"),
             country == "Czechia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Prague"),
             country == "Pakistan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Karachi"),
             country == "China" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Shanghai"),
             country == "Greece" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Athens"),
             country == "Denmark" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Copenhagen"),
             country == "United Arab Emirates" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Dubai"),
             country == "Saudi Arabia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Riyadh"),
             country == "Austria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Vienna"),
             country == "Hungary" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Budapest"),
             country == "Portugal" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Lisbon"),
             country == "Egypt" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Cairo"),
             country == "Bangladesh" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Dhaka"),
             country == "Norway" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Oslo"),
             country == "New Zealand" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Pacific/Auckland"),
             country == "Venezuela" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Caracas"),
             # country == "South Africa" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             country == "Algeria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Algiers"),
             country == "Bulgaria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Sofia"),
             country == "Chile" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Santiago"),
             country == "Morocco" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Casablanca"),
             country == "Serbia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Belgrade"),
             country == "Slovakia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Bratislava"),
             country == "Sri Lanka" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Colombo"),
             country == "Nigeria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Lagos"),
             country == "Croatia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Zagreb"),
             country == "Ecuador" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Guayaquil"),
             country == "Tunisia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Tunis"),
             country == "Belarus" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Minsk"),
             country == "Kazakhstan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Almaty"),
             country == "Finland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Helsinki"),
             country == "Dominican Republic" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Santo_Domingo"),
             country == "Bosnia & Herzegovina" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Sarajevo"),
             country == "Georgia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Tbilisi"),
             country == "Jordan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Amman"),
             country == "Lithuania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Vilnius"),
             country == "Macedonia (FYROM)" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Skopje"),
             country == "Kenya" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Nairobi"),
             country == "Puerto Rico" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Argentina/San_Juan"),
             country == "Slovenia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Ljubljana"),
             country == "Iraq" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Baghdad"),
             country == "Latvia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Riga"),
             country == "Estonia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Tallinn"),
             country == "Uruguay" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Montevideo"),
             country == "Guatemala" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Guatemala"),
             TRUE ~ DateTimeLocal # other value for keeps
           )
         ) %>% 
  mutate(DateTimeLocal = # city
           case_when(
             # city == "Mountain View" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             city == "New York" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/New_York"),
             # city == "San Francisco" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "Sunnyvale" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "San Jose" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "Washington" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             city == "London" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/London"),
             city == "Chicago" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Chicago"),
             city == "Toronto" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Toronto"),
             city == "Jakarta" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jakarta"),
             city == "Melbourne" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Australia/Melbourne"),
             city == "Madrid" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Madrid"),
             # city == "New Delhi" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Santa Clare" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Palo Alto" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Mexico City" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             #  city == "Houston" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Minato" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Seoul" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Atlanta" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Montreal" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Ann Arbor" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             TRUE ~ DateTimeLocal
           )
  ) %>% 
  mutate(DateTimeLocal = DateTimeLocal %>% as.character,
         DateTimeSeattle = DateTimeSeattle %>% as.character) %>% 
  select(-visitStartTime)

# 要記述
test_flat <- 
  test_flat %>% 
    mutate(DateTimeLocal = # Continent
           case_when(
             continent == "Africa" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Tunis"),
             continent == "Americas" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Los_Angeles"),
             continent == "Asia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Bangkok"),
             continent == "Europe" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/London"),
             continent == "Oceania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Australia/Sydney"),
             TRUE ~ DateTimeSeattle
           )
         ) %>% 
  mutate(DateTimeLocal = # country
           case_when(
             country == "United States" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Los_Angeles"),
             country == "India" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Indian/Kerguelen"),
             country == "Canada" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Canada/Atlantic"),
             country == "Japan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Tokyo"),
             country == "Turkey" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Istanbul"),
             country == "Germany" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Berlin"),
             country == "Brazil" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Sao_Paulo"),
             country == "France" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Paris"),
             country == "Mexico" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Mexico_City"),
             country == "Taiwan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Taipei"),
             country == "Russia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Moscow"),
             country == "Spain" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Madrid"),
             country == "Netherlands" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Amsterdam"),
             country == "Italy" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Rome"),
             country == "Poland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Warsaw"),
             country == "Indonesia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jakarta"),
             country == "Philippines" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Manila"),
             country == "Singapore" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Singapore"),
             country == "Ireland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Dublin"),
             country == "Malaysia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Kuala_Lumpur"),
             country == "Romania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Bucharest"),
             country == "Ukraine" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Kiev"),
             country == "Israel" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jerusalem"),
             country == "Peru" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Lima"),
             country == "Sweden" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Stockholm"),
             country == "South Korea" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Seoul"),
             country == "Argentina" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Argentina/Buenos_Aires"),
             country == "Colombia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Bogota"),
             country == "Hong Kong" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Hong_Kong"),
             country == "Belgium" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Brussels"),
             country == "Czechia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Prague"),
             country == "Pakistan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Karachi"),
             country == "China" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Shanghai"),
             country == "Greece" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Athens"),
             country == "Denmark" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Copenhagen"),
             country == "United Arab Emirates" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Dubai"),
             country == "Saudi Arabia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Riyadh"),
             country == "Austria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Vienna"),
             country == "Hungary" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Budapest"),
             country == "Portugal" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Lisbon"),
             country == "Egypt" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Cairo"),
             country == "Bangladesh" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Dhaka"),
             country == "Norway" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Oslo"),
             country == "New Zealand" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Pacific/Auckland"),
             country == "Venezuela" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Caracas"),
             # country == "South Africa" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             country == "Algeria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Algiers"),
             country == "Bulgaria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Sofia"),
             country == "Chile" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Santiago"),
             country == "Morocco" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Casablanca"),
             country == "Serbia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Belgrade"),
             country == "Slovakia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Bratislava"),
             country == "Sri Lanka" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Colombo"),
             country == "Nigeria" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Lagos"),
             country == "Croatia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Zagreb"),
             country == "Ecuador" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Guayaquil"),
             country == "Tunisia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Tunis"),
             country == "Belarus" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Minsk"),
             country == "Kazakhstan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Almaty"),
             country == "Finland" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Helsinki"),
             country == "Dominican Republic" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Santo_Domingo"),
             country == "Bosnia & Herzegovina" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Sarajevo"),
             country == "Georgia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Tbilisi"),
             country == "Jordan" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Amman"),
             country == "Lithuania" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Vilnius"),
             country == "Macedonia (FYROM)" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Skopje"),
             country == "Kenya" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Africa/Nairobi"),
             country == "Puerto Rico" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Argentina/San_Juan"),
             country == "Slovenia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Ljubljana"),
             country == "Iraq" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Baghdad"),
             country == "Latvia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Riga"),
             country == "Estonia" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Tallinn"),
             country == "Uruguay" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Montevideo"),
             country == "Guatemala" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Guatemala"),
             TRUE ~ DateTimeLocal # other value for keeps
           )
         ) %>% 
  mutate(DateTimeLocal = # city
           case_when(
             # city == "Mountain View" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             city == "New York" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/New_York"),
             # city == "San Francisco" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "Sunnyvale" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "San Jose" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             # city == "Washington" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""), # same as los angeles 
             city == "London" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/London"),
             city == "Chicago" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Chicago"),
             city == "Toronto" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "America/Toronto"),
             city == "Jakarta" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Asia/Jakarta"),
             city == "Melbourne" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Australia/Melbourne"),
             city == "Madrid" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = "Europe/Madrid"),
             # city == "New Delhi" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Santa Clare" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Palo Alto" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Mexico City" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             #  city == "Houston" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Minato" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Seoul" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Atlanta" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Montreal" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             # city == "Ann Arbor" ~ as.POSIXct(visitStartTime, origin='1970-01-01', tz = ""),
             TRUE ~ DateTimeLocal
           )
  ) %>% 
  mutate(DateTimeLocal = DateTimeLocal %>% as.character,
         DateTimeSeattle = DateTimeSeattle %>% as.character) %>% 
  select(-visitStartTime)
```

# Export

- 2018/ 9 /27

```{r}
write_csv(train_flat,"~/Desktop/Google_Analytics_Kaggle/input/imp/train_imp1.csv")
write_csv(test_flat,"~/Desktop/Google_Analytics_Kaggle/input/imp/test_imp1.csv")
```
