==> Rmd/2_EDA.Rmd <==
---
title: "Google Analytics Customer: EDA version2"
author: "kotsubo takuto"
output: 
    html_document:
      md_extensions: -ascii_identifiers
      toc: true
      toc_depth: 3
      code_folding: hide
---

# Setting{.tabset .tabset-fade .tabset-pills}

## knitr option

```{r reset, include=FALSE}
# 初期化
rm(list = ls())
```

```{r set up, message=FALSE}
# set directory
setwd("~/Desktop/Google_Analytics_Kaggle/") 
# max.print 
options(max.print="200", digits=5)
# show option  
options("scipen"=100, "digits"=4)
# time zone option
options(readr.default_locale=readr::locale(tz="US/Pacific"))

# Global options
library(knitr)
opts_chunk$set(echo=TRUE,
               cache = FALSE,
	             prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

## Library package

```{r package, message=FALSE}
library(tidyverse)
library(readr) # for csv
library(summarytools) # summary easily for EDA
library(GGally) # ggpairs
library(skimr) 
library(janitor)
library(jsonlite)
library(lubridate)
library(pipeR)
library(scales)
library(gridExtra)
library(ggExtra)
library(countrycode)
library(highcharter)
```

## Load funciton

```{r}
source('~/Desktop/Google_Analytics_Kaggle/script/function.R') # for preprocessing
source('~/Desktop/Google_Analytics_Kaggle/script/makedummies.R') # for preprocessing
```

## Set ggplot theme

```{r}
# set ggplot theme
library(ggthemes)
theme_set(theme_classic(base_size = 14, base_family = "HiraKakuPro-W3"))
```
## Import csv

- Import raw data from `input/imp/{tarin|test}.csv`

```{r cache=TRUE}
# read impute data
train_imp <- read_csv("~/Desktop/Google_Analytics_Kaggle/input/imp/train_imp1.csv") %>% 
  mutate(DateTimeSeattle = DateTimeSeattle %>% as.POSIXct(format = "%Y-%m-%d %H:%M:%S"),
         DateTimeLocal = DateTimeLocal %>% as.POSIXct(format = "%Y-%m-%d %H:%M:%S"))
test_imp <- read_csv("~/Desktop/Google_Analytics_Kaggle/input/imp/test_imp1.csv") %>% 
  mutate(DateTimeSeattle = DateTimeSeattle %>% as.POSIXct(format = "%Y-%m-%d %H:%M:%S"),
         DateTimeLocal = DateTimeLocal %>% as.POSIXct(format = "%Y-%m-%d %H:%M:%S"))
```

# Explatory Data Analysis

- Apply glimpse and dfsummary to the all data
- Check three id values
- Confirm relationship with target variables: not yet

## Apply glimpse and dfsummary to the all data{.tabset .tabset-fade .tabset-pills}

### glimpse 

```{r cache=TRUE}
glimpse(train_imp)
glimpse(test_imp)
```

### dfsummary

```{r cache=TRUE}
dfSummary(train_imp) %>% 
  view(method = "render")
dfSummary(test_imp) %>% 
  view(method = "render")
```

## Check three id values

それぞれのidが何を意味しているのか, 集約する上で理解する必要がある. 前処理したデータに対して, あらためて調査する.

- `fullVisitorId`
- `visitId`

各idにおいて同値のidが何を意味するのか調べる. 

- `fullVisitorId`:
    - user idと考えて良さそう, 個人情報が同じ. 
    - 各user の`newVisits`が複数回ある場合がある (2: 2834, 3:30). 
    - 各userの`visitNumber`にも同じ値が出現していることある (2:4052, 3: 42).
    - 解決策: 観測期間でid振り直して, id == 1 & visitNumber == 1の行を newVisits =1とする?  

```{r cache=TRUE}
train_imp %>% 
  group_by(fullVisitorId) %>% 
  mutate(num = row_number()) %>% 
  filter(num == max(num)) %>% 
  ungroup() %>% 
  select(fullVisitorId,num) -> tmp

tmp %>% 
  inner_join(train_imp,by=c("fullVisitorId")) -> tmp1

train_imp %>% 
  filter(newVisits == 1) %>% 
  group_by(fullVisitorId) %>% 
  mutate(num1 = n()) %>% 
  tabyl(num1)

train_imp %>% 
  group_by(fullVisitorId, visitNumber) %>% 
  mutate(num = n()) %>% 
  ungroup() %>% 
  tabyl(num)
```

- `visitId`:
    - システム側の id, visitStartTimeが同じであることから推測
    - `visitId`と`DateTimeSeattle`のuniqueな個数が一致しない (dif 856個) -> user 側の time zone 使われている??
    - dateと比べて情報ないので不要??

```{r cache=TRUE}
train_imp %>% 
  group_by(visitId) %>% 
  mutate(num = row_number()) %>% 
  filter(num == max(num)) %>% 
  ungroup() %>% 
  select(visitId,num) -> tmp

tmp %>% 
  inner_join(train_imp,by=c("visitId")) -> tmp2

train_imp %>>%
  (~ distinct(.,visitId) %>% NROW %>% print) %>% 
  distinct(DateTimeSeattle) %>% NROW
```

- `fullVisitorId` & `visitId` : 
    - 重複 898件, `visitNumber`は一致する
    - 日付情報や, userの情報が異なる場合ある

```{r cache=TRUE}
train_imp %>% 
  group_by(fullVisitorId, visitId) %>% 
  mutate(num = row_number()) %>% 
  filter(num == max(num)) %>% 
  ungroup() %>% 
  select(fullVisitorId,visitId,num) -> tmp

tmp %>% 
  inner_join(train_imp,by=c("fullVisitorId","visitId")) -> tmp3

tmp3 %>% 
  filter(num == 2) %>% 
  select(fullVisitorId, visitId,visitNumber) %>% 
  group_by(fullVisitorId,visitId) %>% 
  mutate( Difference = visitNumber - lag(visitNumber)) %>% 
  tabyl(Difference)
```

## Confirm relationship with target variables

### Chech target values

- 0より大きいのtotalRevenueに対して, log1p()を計算.

```{r cache=TRUE}
train_imp %>% 
  group_by(fullVisitorId) %>% 
  summarise(rev_sum = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  filter(rev_sum > 0) %>% 
  ggplot(aes(x=log1p(rev_sum), y=..density..)) + 
  geom_histogram(fill='steelblue', na.rm=TRUE, bins=40) + 
  geom_density(aes(x=log1p(rev_sum)), fill='orange', color='orange', alpha=0.3, na.rm=TRUE) + 
  labs(
    title = 'Distribution of sum of transaction revenue without 0',
    x = 'Natural log1p of transaction revenue'
  )
```

### day & target

- Seattle Time

```{r fig.width=10, fig.height=10, cache=TRUE}
## daily visit in train data
g1 <- 
  train_imp %>%
  mutate(date = DateTimeSeattle %>% date) %>% 
  group_by(date) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=date, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='Visits number',
    title='Daily visits in train data'
  ) +
  scale_x_date(labels = date_format("%Y-%m"))
## daily visit in test data
g2 <- 
  test_imp %>%
  mutate(date = DateTimeSeattle %>% date) %>% 
  group_by(date) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=date, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='Visits number',
    title='Daily visits in test data'
  ) +
  scale_x_date(labels = date_format("%Y-%m"))
## log(tatalRevenue) in train data
g3 <-
  train_imp %>%
  mutate(date = DateTimeSeattle %>% date) %>% 
  group_by(date) %>% 
  summarise(day_rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  ggplot(aes(x=date, y=day_rev)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='totalRevenue',
    title='Daily transaction revenue'
  ) +
  scale_x_date(labels = date_format("%Y-%m")) +
  scale_y_continuous(labels = scales::scientific)

grid.arrange(g1, g2, g3, nrow=3)
```

- Local Time

```{r fig.width=10, fig.height=10, cache=TRUE}
## daily visit in train data
g1 <- 
  train_imp %>%
  mutate(date = DateTimeLocal %>% date) %>% 
  group_by(date) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=date, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='Visits number',
    title='Daily visits in train data'
  ) +
  scale_x_date(labels = date_format("%Y-%m"))
## daily visit in test data
g2 <- 
  test_imp %>%
  mutate(date = DateTimeLocal %>% date) %>% 
  group_by(date) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=date, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='Visits number',
    title='Daily visits in test data'
  ) +
  scale_x_date(labels = date_format("%Y-%m"))
## log(tatalRevenue) in train data
g3 <-
  train_imp %>%
  mutate(date = DateTimeLocal %>% date) %>% 
  group_by(date) %>% 
  summarise(day_rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  ggplot(aes(x=date, y=day_rev)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='',
    y='totalRevenue',
    title='Daily transaction revenue'
  ) +
  scale_x_date(labels = date_format("%Y-%m")) +
  scale_y_continuous(labels = scales::scientific)

grid.arrange(g1, g2, g3, nrow=3)
```
### hour & target

- Seattle Time

```{r fig.width=10, fig.height=10, cache=TRUE}
## daily visit in train data
g1 <- 
  train_imp %>%
  mutate(hour = DateTimeSeattle %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='Visits number',
    title='visits number in train data by hour of day'
  )

## daily visit in test data
g2 <- 
  test_imp %>%
  mutate(hour = DateTimeSeattle %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='Visits number',
    title='visits number in test data by hour of day'
  ) 

## log(tatalRevenue) in train data
g3 <-
  train_imp %>% 
  mutate(hour = DateTimeSeattle %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_rev)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='totalRevenue',
    title='Transaction revenue by hour of day'
  ) +
  scale_y_continuous(labels = scales::scientific)

grid.arrange(g1, g2, g3, nrow=3)
```

- Local Time

```{r fig.width=10, fig.height=10, cache=TRUE}
## daily visit in train data
g1 <- 
  train_imp %>%
  mutate(hour = DateTimeLocal %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='Visits number',
    title='visits number in train data by hour of day'
  )

## daily visit in test data
g2 <- 
  test_imp %>%
  mutate(hour = DateTimeLocal %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_count)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='Visits number',
    title='visits number in test data by hour of day'
  ) 

## log(tatalRevenue) in train data
g3 <-
  train_imp %>% 
  mutate(hour = DateTimeLocal %>% hour) %>% 
  group_by(hour) %>% 
  summarise(day_rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  ggplot(aes(x=hour, y=day_rev)) + 
  geom_line(color='steelblue') +
  geom_smooth(color='orange', method = 'loess', fomula = 'y ~ x') + 
  labs(
    x='Hour of day',
    y='totalRevenue',
    title='Transaction revenue by hour of day'
  ) +
  scale_y_continuous(labels = scales::scientific)

grid.arrange(g1, g2, g3, nrow=3)
```

### Device & target

- `browser`: top 9種に0より大きいrevenue.
- `operatingSystem`: top 7個に0より大きいrevenue.
- `isMobile`: deviceCategoryの情報に集約されているので不要??
- `deviceCategory`: 3種

- 少しFirefoxがつよいのかなというイメージ

```{r fig.width=20, fig.height=20, cache=TRUE}
g1 <-
  train_imp %>% 
  group_by(browser) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(15, count) %>% 
  ggplot(aes(x=reorder(browser, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y='Access number',
       title='Distribution of device browser') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(browser) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(9, rev) %>% 
  ggplot(aes(x=reorder(browser, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='browser',
       y='totalRevenue',
       title='Distribution of revenue by device browser') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=7) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

- WindowsとMacintoshの順位が変わっている. 

```{r fig.width=20, fig.height=20, cache=TRUE}
g1 <-
  train_imp %>% 
  group_by(operatingSystem) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(15, count) %>% 
  ggplot(aes(x=reorder(operatingSystem, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y='Access number',
       title='Distribution of device operatingSystem') +
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(operatingSystem) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(7, rev) %>% 
  ggplot(aes(x=reorder(operatingSystem, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='operatingSystem',
       y='totalRevenue',
       title='Distribution of revenue by device operatingSystem') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=8) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

- 全体比で見ると, mobileの割合が少ないかなという感じ.

```{r fig.width=20, fig.height=20, cache=TRUE}
g1 <-
  train_imp %>% 
  group_by(deviceCategory) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  ggplot(aes(x=reorder(deviceCategory, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y='Access number',
       title='Distribution of deviceCategory') +
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(deviceCategory) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  ggplot(aes(x=reorder(deviceCategory, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='deviceCategory',
       y='totalRevenue',
       title='Distribution of revenue by deviceCategory') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=8) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, ncol=2)
```

- pipeRで遊ぶ, 全く関係ない

```{r eval=FALSE}
train_imp %>>% 
  (~ g1 = 
     group_by(.,deviceCategory) %>>% 
     summarise(count = n()) %>>%
     ungroup() %>>% 
     ggplot(aes(x=reorder(deviceCategory, -count), y=count)) +
     geom_bar(stat='identity', fill='steelblue') +
     labs(x='',
          y='Access number',
          title='Distribution of deviceCategory') +
     geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=4) +
     # scale_y_continuous(labels = scales::scientific) +
     theme(axis.text.x = element_text(angle = 30, hjust = 1))
   ) %>>%
  summarise(rev = sum(transactionRevenue)) %>>% 
  ungroup() %>>% 
  ggplot(aes(x=reorder(deviceCategory, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='deviceCategory',
       y='totalRevenue',
       title='Distribution of revenue by deviceCategory') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=4) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1)) -> g2
grid.arrange(g1, g2, ncol=2)
```

### geoNetwork & target

- `continent`: 6種類
- `subContinent`: 上位10種類
- `country`: 上位10種類
- `region`: 上位10種類
- `metro`: 上位10種類
- `city`: 上位10種類
- `networkDomain`: 上位10種類

#### continent 

- 比率でみると, Africaが強い, Europeはアクセス数の割に購入金額が少ない

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(continent) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  # top_n(15, count) %>% 
  ggplot(aes(x=reorder(continent, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of continent') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(continent) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  # top_n(9, rev) %>% 
  ggplot(aes(x=reorder(continent, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='continent',
       y='totalRevenue',
       title='Distribution of revenue by continent') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=8) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### sub continent

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(subContinent) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(subContinent, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of Sub continent') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(subContinent) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(subContinent, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='subcontinent',
       y='totalRevenue',
       title='Distribution of revenue by Sub continent') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### country

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(country) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(country, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of country') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(country) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(country, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='country',
       y='totalRevenue',
       title='Distribution of revenue by country') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### region

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(region) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(region, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of region') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(region) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(region, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='region',
       y='totalRevenue',
       title='Distribution of revenue by region') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### metro 

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(metro) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(metro, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of metro') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(metro) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(metro, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='metro',
       y='totalRevenue',
       title='Distribution of revenue by metro') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### city

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(city) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(city, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of city') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(city) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(city, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='city',
       y='totalRevenue',
       title='Distribution of revenue by city') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

#### networkDomain

```{r fig.width=20, fig.height=20}
g1 <-
  train_imp %>% 
  group_by(networkDomain) %>% 
  summarise(count = n()) %>%
  ungroup() %>% 
  top_n(10, count) %>% 
  ggplot(aes(x=reorder(networkDomain, -count), y=count)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='',
       y= 'Access number',
       title='Distribution of networkDomain') +  
  geom_text(aes(label=count), position = position_stack(vjust = 0.5),size=8) +
  # scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

g2 <-
  train_imp %>% 
  group_by(networkDomain) %>% 
  summarise(rev = sum(transactionRevenue)) %>% 
  ungroup() %>% 
  top_n(10, rev) %>% 
  ggplot(aes(x=reorder(networkDomain, -rev), y=rev)) +
  geom_bar(stat='identity', fill='steelblue') +
  labs(x='networkDomain',
       y='totalRevenue',
       title='Distribution of revenue by networkDomain') +
  geom_text(aes(label=format(rev,big.mark=",",scientific=FALSE)), position = position_stack(vjust = 0.5),size=6) +
  scale_y_continuous(labels = scales::scientific) +
  theme(axis.text.x = element_text(angle = 30, hjust = 1))

grid.arrange(g1, g2, nrow=2)
```

### geo plot

```{r}
tmp <- train_imp %>% 
  group_by(country) %>% 
  summarise(rev = sum(transactionRevenue)/ 10^6) %>% 
  ungroup() %>% 
  mutate(iso3 = countrycode(country,origin='country.name', destination='iso3c'))
  
  
highchart() %>% 
  hc_add_series_map(worldgeojson,tmp, value = 'rev', joinBy = 'iso3') %>%
    hc_title(text = 'Total transaction revenue by country (divided by 10^6)') %>%
    hc_subtitle(text = "August 2016 to August 2017") %>%
    hc_tooltip(useHTML = TRUE, headerFormat = "",pointFormat = "{point.country}: ${point.rev:.0f}")
```

### Total Features

```{r fig.width=20, fig.height=20}
g1 <- ggplot(train_imp, aes(x=log(pageviews), y=log1p(transactionRevenue))) + 
  geom_point(color='steelblue') +
  geom_smooth(method='lm', color='orange') + 
  labs(
    y='Transaction revenue (log)',
    title='Pageviews vs transaction revenue',
    subtitle='visit-level')
  

g2 <- ggplot(train_imp, aes(x=log(hits), y=log1p(transactionRevenue))) + 
  geom_point(color='steelblue') +
  geom_smooth(method='lm', color='orange') + 
  labs(
    y='Transaction revenue (log)',
    title='Hits vs transaction revenue',
    subtitle='visit-level')

m1 <- ggMarginal(g1, type='histogram', fill='steelblue')
m2 <- ggMarginal(g2, type='histogram', fill='steelblue')

grid.arrange(m1, m2, nrow = 1, ncol = 2)
```
